FVTA with Phong shading
Sorry guys, there is something wrong with github manager in VS2022, and I cann't push and merge the code in my VS2022. I am still trying to fix it. So as not to disrupt everyone's progress, I'm posting my update on FVTA here. The mian difference is all in fvta_step(). Don't forget to add uniform vec3 light_direction, which is currently glm::vec3(0.2f,0.2f,0.2f). Sorry for that once again.